+++ functions/src/services/MaintainService.ts
@@ -1,6 +1,8 @@
 import { SheetService } from "./SheetService"
 import { maintainColumn } from "../models/sheetColumn"
 import { Maintain } from "../models/Maintain";
+import {LineBotService} from "./LineBotService";
+import {TemplateMessage, TextMessage} from "@line/bot-sdk";
 
 export class MaintainService {
 
@@ -25,4 +27,59 @@ export class MaintainService {
 
         return maintain;
     }
+
+    public getMaintainState(maintainState: string): string {
+        if (maintainState == "0"){
+            return "尚未完成"
+        }else if (maintainState =="1"){
+            return "已在維修中，請耐心等候"
+        }else{
+            return "已經完成囉，感謝您的報修"
+        }
+    };
+
+    public requestReport(userId: string, result: any): Promise<any> {
+        let url = `https://docs.google.com/forms/d/e/1FAIpQLSd_xr_18k4FIPjBYECcwv2fc1dOT_IuZMxAgGJUuseg9KInmw/viewform?usp=pp_url&entry.815484785&entry.534784453&entry.1173029400&entry.142495844&entry.1574186958&entry.780437475=${userId}`;
+        const lineBotService = new LineBotService();
+        const lineMessage: TemplateMessage = {
+            type: "template",
+            altText: "This is a buttons template",
+            template: {
+                type: "buttons",
+                title: "維修系統報修",
+                text: "請填寫報修資料",
+                actions: [
+                    {
+                        type: "uri",
+                        label: "點擊填表",
+                        uri: url
+                    }
+                ]
+            }
+        };
+
+        return lineBotService.pushMessage(userId, lineMessage)
+    };
+
+    public async searchReport (userId: string, result: any) {
+        const maintain = await this.getMaintainById(result.parameters.number);
+        const lineBotService = new LineBotService();
+        if (maintain.Id == null){
+            const lineMessage: TextMessage = {
+                type: "text",
+                text: `您所查詢的單號不存在`
+            };
+
+            lineBotService.pushMessage(userId, lineMessage)
+        }else {
+            const maintainState = this.getMaintainState(maintain.MaintainState);
+
+            const lineMessage: TextMessage = {
+                type: "text",
+                text: `單號：${maintain.Id}\n${maintain.Locate}樓 ${maintain.Item}\n目前的維修狀態為${maintainState}`
+            };
+
+            lineBotService.pushMessage(userId, lineMessage)
+        }
+    };
 }
